Version .3
Lots of changes...

Menu is working, animated, Font's have been chosen
Backend is designed in c# and app can now effectively communicate using RSA encoded json request/response system
This means we can implement log in securely! (Hopefully Securely)

Currently working on te Game Over Activity, this is the screen that the game shows after a game.
The retry button is not coded yet, and not every label has the data linked.
Data is transmitted from the getFinalStats  method in the Logic Engine

The best way to hop into this commit is by checking all the comments!
KNOW what you can do while there is currently no documentation by method checking the Components

I've done my best to keep clear consistant comments in places I expect will be needed...
Of course I'm not perfect and my intuition is not like everyone, so there may be a place where I've not left
clear enough documentation in-code. if so, don't hesitate to contact me....

Apologies for lack of indepth covergage, It is currently 4 in the morning!